aibuddy/requirements.txt
botbuilder-core==4.14.8
botbuilder-schema==4.14.8
botbuilder-integration-aiohttp==4.14.8
aiohttp==3.9.1
groq==0.11.0
python-dotenv==1.0.0
